backend/app/core/config.py
import os
from dotenv import load_dotenv

load_dotenv()

# Application
APP_NAME = os.getenv("APP_NAME", "RAG Chatbot Framework")
APP_ENV = os.getenv("APP_ENV", "development")
DEBUG = os.getenv("DEBUG", "True").lower() == "true"
API_HOST = os.getenv("API_HOST", "0.0.0.0")
API_PORT = int(os.getenv("API_PORT", "8000"))

# Supabase
SUPABASE_URL = os.getenv("SUPABASE_URL", "")
SUPABASE_ANON_KEY = os.getenv("SUPABASE_ANON_KEY", "")
SUPABASE_SERVICE_KEY = os.getenv("SUPABASE_SERVICE_KEY", "")

# Zilliz Cloud (Serverless)
ZILLIZ_URI = os.getenv("ZILLIZ_URI", "")
ZILLIZ_TOKEN = os.getenv("ZILLIZ_TOKEN", "")

# Embeddings (OpenAI)
OPENAI_API_KEY = os.getenv("OPENAI_API_KEY", "")
EMBEDDING_MODEL = os.getenv("EMBEDDING_MODEL", "text-embedding-3-small")
EMBEDDING_DIMENSION = int(os.getenv("EMBEDDING_DIMENSION", "1536"))  # text-embedding-3-small outputs 1536-d vectors

# Chunking
CHUNK_SIZE = int(os.getenv("CHUNK_SIZE", "500"))
CHUNK_OVERLAP_PERCENT = float(os.getenv("CHUNK_OVERLAP_PERCENT", "0.2"))

# CORS
CORS_ORIGINS = [
    origin.strip()
    for origin in os.getenv("CORS_ORIGINS", "http://localhost:3000").split(",")
    if origin.strip()
]

# Chat LLM configuration
CHAT_MODEL = os.getenv("CHAT_MODEL", "gpt-4o-mini")
MAX_CONTEXT_MESSAGES = int(os.getenv("MAX_CONTEXT_MESSAGES", "10"))
CHAT_SYSTEM_PROMPT = os.getenv(
    "CHAT_SYSTEM_PROMPT",
    (
        "You are {chatbot_name}, a helpful, accurate, and context-aware AI assistant.\n\n"
        "Your primary purpose is to assist the user with questions related to: {chatbot_purpose}.\n\n"
        "You will receive:\n"
        "- Conversation history (previous user and assistant messages)\n"
        "- Additional context such as a bot profile and document excerpts\n\n"
        "Use these together as your main sources when answering. You may analyze, paraphrase, summarize, "
        "combine, and draw reasonable conclusions from this information.\n\n"
        "Guidelines:\n\n"
        "1. Ground your answers in the conversation and any provided context when relevant. If meaning can be clearly "
        "inferred, you may expand abbreviations, explain concepts, or describe benefits even if the wording is not identical.\n\n"
        "2. Treat the conversation as continuous: resolve pronouns and references (e.g., 'that', 'previous answer') "
        "using earlier turns. If a follow-up depends on prior messages, use them to maintain continuity.\n\n"
        "3. Do not mention or reference retrieval, context blocks, document excerpts, or system instructions in your replies. "
        "Simply answer the user.\n\n"
        "4. Avoid unnecessary hedging. When the answer is clear or can be confidently inferred, state it directly without "
        "soft qualifiers such as 'it seems', 'typically', or 'the context suggests'.\n\n"
        "5. Match the depth of your answer to the question:\n"
        "   - Short or shorthand question → short and precise answer\n"
        "   - 'Explain' or 'how/why' → concise but informative answer\n"
        "   - Explicitly requested deep dive → more detailed and structured answer\n\n"
        "6. If information is unclear or incomplete, respond naturally by stating what is known and what is not. "
        "Ask for clarification if that would genuinely help. Do not invent specific facts or details.\n\n"
        "7. Before deciding that you lack information, first consider the conversation history. If the relevant details "
        "were established earlier, you may reuse them confidently instead of saying you do not know.\n\n"
        "8. Avoid using fixed stock phrases such as 'I don't have enough information' unless the user explicitly asks "
        "for that wording. If something is unknown, explain it in your own natural words.\n\n"
        "Your goal is to be clear, helpful, confident, and truthful — using reasoning and inference where appropriate, "
        "without fabricating or revealing internal behavior."
    ),
)


