backend/requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database & Storage
supabase==2.0.3

# Vector Database - Zilliz Cloud (Milvus)
pymilvus==2.3.4

# AI & Embeddings
openai==1.12.0
numpy<2.0.0

# Document Processing
pypdf==3.17.1
python-docx==1.1.0
openpyxl==3.1.2
python-magic==0.4.27
markdown==3.5.1
langchain==0.1.0
langchain-community==0.0.20

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Utilities
httpx>=0.24.0,<0.25.0
aiofiles==23.2.1
python-dateutil==2.8.2
tiktoken==0.5.2
marshmallow<4.0.0,>=3.0.0

# Firebase (for later - chat history)
firebase-admin==6.3.0

